trying to make progress on getting to better structure, impml
